Macro test_exec::exec

macro_rules! exec {
    (
        // Input Configuration
        $program:expr // the program name, without a key and required
        $(, args: $args:expr)? // the arguments to pass to the program
        $(, cwd: $cwd:expr)? // the desired current working directory
        $(, clear_env: $clear_env:expr)? // a boolean setting whether the environment should be cleared
        $(, env: { $( $key:ident : $value:expr ),* } )? // the environment variables as pseudo-object
        $(, modify_path: $modify_path:expr)? // enable the auto-modification of the PATH to include bin targets
        $(, stdin: $stdin:expr)? // what to write to the program's stdin
        $(, timeout: $timeout:expr)? // maximum allowed running time of the program
        $(, log: $log:expr)? // log the output

        // Assertions
        $(, code: $code:expr)? // the expected exit code
        $(, stdout: $stdout:expr)? // the expected stdout
        $(, stderr: $stderr:expr)? // the expected stderr
        $(, signal: $signal:expr)? // expected signal ID of the signal that terminated the program
    ) => { ... };
    ($program:expr,) => { ... };
}

Comfortably execute a command and its output.

This macro supports a variety of input.

Input configuration

args (impl IntoIter<Item = impl AsRef<OsStr>>)

Configure the arguments to pass to the executable. Array literals and vec![] declarations are possible.

exec!{
    "echo",
    args: ["Hello world!"]
};

cwd (impl AsRef<OsStr>)

Set the current working directory of the program.

#[cfg(unix)]
exec!{
    "pwd",
    cwd: "/tmp"
};

#[cfg(windows)]
exec!{
    "cd",
    cwd: "/system32"
};

env and clear_env (bool)

Modify enviroment variables. env is a pseudo-object with raw identifiers as keys and expressions where impl AsRef<OsStr> as values, separated by commas.

The clear_env option clears all environment variables before applying the env values. It is set to false by default. Remember to keep the path of the program in the path when using this key.

exec!{
    "printenv",
    clear_env: true,
    env: {
        USER: "root",
        COFFEE: "a lot"
    }
};

modify_path (bool)

By default, this is set to true, so the target/debug and target/release directorys are added to the PATH automatically. This modification also happens if a custom PATH is provided via the env key.

If this behavior is undesired, this key can be set to false.

exec!{
    "printenv",
    args: ["PATH"],
    env: {
        PATH: "/bin"
    },
    modify_path: false,

    stdout: b"/bin\n"
};

stdin (impl AsRef<[u8]>)

Set data to write to the program's stdin.

exec!{
    "cat",
    stdin: b"meow"
};

timeout (u64)

Set the maximum running time for the program in milliseconds. When the timeout is exceeded, the process is killed using SIGKILL on Unix. The exit code is set to the signal number, which is 9 on Unix. This exit code is often unexpected, so use this carefully and only when the test if supposed to fail after a certain time.

The maximum time depends on the operating system, being i32::max_value() on Unix and u32::max_value() on Windows (no warranty).

If timeout is not given, the program will never timeout.

exec!{
    "sleep",
    args: ["60"],
    timeout: 5000 // milliseconds!
};

log (bool)

By setting this to true, the output of the program is logged after successful execution.

Remember to pass --nocapture to tests using this option.

Output comparison

exec offers various ways to compare the output of the program directly through the macro. But, for dynamic output, the Output struct is returned.

let output = exec!("pwd");
assert_eq!(output, b"/");

code (i32)

Make sure the exit code equals this code, panic if otherwise. This property is always compared, and the default is 0.

exec!{
    "false",
    code: 1
};
exec!{
    "true",
    code: 0 // unnecessary, but may be used for more explicitness
};

stdout and stderr (impl AsRef<[u8]>)

Make sure the program's stdout and stderr are exactly these bytes. This is not compared if omitted. Keep in my mind that many programs append a \n to their stdout.

exec!{
    "echo",
    args: ["Hello", "world"],

    stdout: b"Hello world\n",
    stderr: []
};

signal (i32)

! Supported Unix only, ignored on Windows

Make sure the signal that terminated the program equals this signal. It is required that the program was definitely stopped by this signal, otherwise the assertion will fail.

exec!{
    "sleep",
    args: ["60"],
    timeout: 1000,

    code: 9,
    signal: 9 // SIGKILL = 9
};